Autocomplete: Extract Agent tests
