+++ agent/src/__tests__/autocomplete/src/sum.ts
@@ -1,0 +1,3 @@
+export function sum(a: number, b: number): number {
+    /* CURSOR */
+}
